requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.2
openai==2.31.0
redis==5.2.1
reportlab==4.4.1
python-docx==1.1.2